Move renderer to Infinite Scroll component
